requirements.txt
scipy
numpy==1.22.0
matplotlib
opencv_python
torch
torchvision
torchmetrics
tqdm
Pillow
h5py
tkintertable
scikit-image
scikit-learn
colormath==3.0.0
pandas
openpyxl
chardet
PyQt5
imutils